+++ src/main.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::large_enum_variant, clippy::too_many_arguments)]
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
+mod appearance;
 mod audio;
 mod buffer;
 mod event;
@@ -185,7 +186,7 @@ impl Halloy {
             Halloy {
                 version: Version::new(),
                 screen,
-                theme: config.themes.default.clone().into(),
+                theme: appearance::theme(&config.appearance.selected).into(),
                 clients: Default::default(),
                 servers: config.servers.clone(),
                 config,
@@ -207,7 +208,7 @@ pub enum Screen {
 
 #[derive(Debug)]
 pub enum Message {
-    ThemesReloaded(config::Themes),
+    AppearanceReloaded(data::appearance::Appearance),
     ScreenConfigReloaded(Result<Config, config::Error>),
     Dashboard(dashboard::Message),
     Stream(stream::Update),
@@ -219,6 +220,7 @@ pub enum Message {
     Version(Option<String>),
     Modal(modal::Message),
     RouteReceived(String),
+    AppearanceChange(appearance::Mode),
     Window(window::Id, window::Event),
     WindowSettingsSaved(Result<(), window::Error>),
 }
@@ -283,8 +285,8 @@ impl Halloy {
 
     fn update(&mut self, message: Message) -> Task<Message> {
         match message {
-            Message::ThemesReloaded(updated) => {
-                self.config.themes = updated;
+            Message::AppearanceReloaded(appearance) => {
+                self.config.appearance = appearance;
                 Task::none()
             }
             Message::ScreenConfigReloaded(updated) => {
@@ -321,7 +323,7 @@ impl Halloy {
                                     .collect::<Vec<_>>();
 
                                 self.servers = updated.servers.clone();
-                                self.theme = updated.themes.default.clone().into();
+                                self.theme = appearance::theme(&updated.appearance.selected).into();
                                 self.config = updated;
 
                                 for server in removed_servers {
@@ -335,8 +337,8 @@ impl Halloy {
                         Task::none()
                     }
                     Some(dashboard::Event::ReloadThemes) => Task::future(Config::load())
-                        .and_then(|config| Task::done(config.themes))
-                        .map(Message::ThemesReloaded),
+                        .and_then(|config| Task::done(config.appearance))
+                        .map(Message::AppearanceReloaded),
                     Some(dashboard::Event::QuitServer(server)) => {
                         self.clients.quit(&server, None);
                         Task::none()
@@ -882,6 +884,18 @@ impl Halloy {
 
                 Task::none()
             }
+            Message::AppearanceChange(mode) => {
+                if let data::appearance::Selected::Dynamic { light, dark } =
+                    &self.config.appearance.selected
+                {
+                    self.theme = match mode {
+                        appearance::Mode::Dark => dark.clone().into(),
+                        appearance::Mode::Light => light.clone().into(),
+                    }
+                }
+
+                Task::none()
+            }
         }
     }
 
@@ -964,6 +978,7 @@ impl Halloy {
             url::listen().map(Message::RouteReceived),
             events().map(|(window, event)| Message::Event(window, event)),
             window::events().map(|(window, event)| Message::Window(window, event)),
+            appearance::subscription().map(Message::AppearanceChange),
             tick,
             streams,
         ])